backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "librosa>=0.11.0",
    "mido>=1.3.3",
    "pydantic-ai>=1.52.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
